requirements.txt
asgiref==3.8.1
click==8.2.1
dj-database-url==3.0.0
Django==5.2.3
django-async-sse==1.0.0
djangorestframework==3.16.0
h11==0.16.0
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
uvicorn==0.34.3
